+++ src/DefaultTestOptions.spec.ts
@@ -14,7 +14,6 @@ export const DefaultTestOptions = {
   // onIdleIntervalMillis: xxx
   maxTasksPerProcess: 5,
   taskTimeoutMillis: 250,
-  maxReasonableProcessFailuresPerMinute: 0, // disable. We're expecting flakiness.
   // we shouldn't need these overrides...
   // ...(isCI ? { streamFlushMillis: bco.streamFlushMillis * 3 } : {}),
   // onIdleIntervalMillis: 1000,